Rename bindings files to match the Python package name
